RAG/requirements_for_RAG.txt
faiss-cpu
sentence-transformers
mistralai
numpy
python-dotenv
transformers
torch
sounddevice
scipy
pyttsx3
gTTS
google-genai